Final fixes.

+++ Heleonix.Build/Tasks/OpenCover.cs
@@ -41,22 +41,22 @@ namespace Heleonix.Build.Tasks
         /// <summary>
         /// Gets or sets the minimum class coverage, in range: 0% - 100%.
         /// </summary>
-        public float MinClassCoverage { get; set; }
+        public int MinClassCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum method coverage, in range: 0% - 100%.
         /// </summary>
-        public float MinMethodCoverage { get; set; }
+        public int MinMethodCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum branch coverage, in range: 0% - 100%.
         /// </summary>
-        public float MinBranchCoverage { get; set; }
+        public int MinBranchCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum line coverage, in range: 0% - 100%.
         /// </summary>
-        public float MinLineCoverage { get; set; }
+        public int MinLineCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the filters to exclude code from coverage by attribute in format: Name*;*Attribute.
@@ -163,25 +163,25 @@ namespace Heleonix.Build.Tasks
         /// Gets or sets the class coverage, in range: 0% - 100% [Output].
         /// </summary>
         [Output]
-        public float ClassCoverage { get; set; }
+        public int ClassCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the method coverage, in range: 0% - 100% [Output].
         /// </summary>
         [Output]
-        public float MethodCoverage { get; set; }
+        public int MethodCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the line coverage, in range: 0% - 100% [Output].
         /// </summary>
         [Output]
-        public float LineCoverage { get; set; }
+        public int LineCoverage { get; set; }
 
         /// <summary>
         /// Gets or sets the branch coverage, in range: 0% - 100% [Output].
         /// </summary>
         [Output]
-        public float BranchCoverage { get; set; }
+        public int BranchCoverage { get; set; }
 
         /// <summary>
         /// Executes the OpenCover.
@@ -275,12 +275,12 @@ namespace Heleonix.Build.Tasks
             this.MaxCyclomaticComplexity = Convert.ToInt32(
                 summary.Attribute("maxCyclomaticComplexity").Value,
                 NumberFormatInfo.InvariantInfo);
-            this.ClassCoverage = (float)this.VisitedClasses / this.TotalClasses * 100;
-            this.MethodCoverage = (float)this.VisitedMethods / this.TotalMethods * 100;
-            this.LineCoverage = Convert.ToSingle(
+            this.ClassCoverage = (int)((float)this.VisitedClasses / this.TotalClasses * 100);
+            this.MethodCoverage = (int)((float)this.VisitedMethods / this.TotalMethods * 100);
+            this.LineCoverage = (int)Convert.ToSingle(
                 summary.Attribute("sequenceCoverage").Value,
                 NumberFormatInfo.InvariantInfo);
-            this.BranchCoverage = Convert.ToSingle(
+            this.BranchCoverage = (int)Convert.ToSingle(
                 summary.Attribute("branchCoverage").Value,
                 NumberFormatInfo.InvariantInfo);
